Pass data type when writing meshtag
